django_remote/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["bafaurazan <rafalbe777@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
django = "^5.0.4"
django-ninja = "^1.1.0"
black = "^24.4.0"
ruff = "^0.4.1"
pexpect = "^4.9.0"
matplotlib = "^3.9.2"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
